3. Время разбивать камни_корректировка_2
