+++ src/view/films-list.js
@@ -1,22 +1,15 @@
-const createFilmsSectionTemplate = () => {
-  return `<section class="films">
-
-</section>
-  `;
-};
 
 const createFilmsListTemplate = (listType, additionalClass = ``) => {
-  return `<section class="films-list ${additionalClass}">
-  <h2 class="films-list__title">${listType}</h2>
+  return (
+    `<section class="films-list ${additionalClass}">
+      <h2 class="films-list__title">${listType}</h2>
 
-  <div class="films-list__container">
+       <div class="films-list__container">
 
-  </div>
-</section>`;
+       </div>
+    </section>`
+  );
 };
 
 
-export {
-  createFilmsSectionTemplate,
-  createFilmsListTemplate
-};
+export {createFilmsListTemplate};
