+++ src/view/films-wrapper.js
@@ -1,0 +1,9 @@
+const createFilmsSectionTemplate = () => {
+  return (
+    `<section class="films">
+
+     </section>`
+  );
+};
+
+export {createFilmsSectionTemplate};
