+++ src/view/footer-stats.js
@@ -1,0 +1,5 @@
+const createFilmsStatistics = () => {
+  return `<p>130 291 movies inside</p>`;
+};
+
+export {createFilmsStatistics};
