+++ src/view/movie-card.js
@@ -1,5 +1,6 @@
 const createFilmArticle = () => {
-  return `<article class="film-card">
+  return (
+    `<article class="film-card">
         <h3 class="film-card__title">Popeye the Sailor Meets Sindbad the Sailor</h3>
         <p class="film-card__rating">6.3</p>
         <p class="film-card__info">
@@ -15,9 +16,8 @@ const createFilmArticle = () => {
           <button class="film-card__controls-item button film-card__controls-item--mark-as-watched film-card__controls-item--active" type="button">Mark as watched</button>
           <button class="film-card__controls-item button film-card__controls-item--favorite film-card__controls-item--active" type="button">Mark as favorite</button>
         </div>
-      </article>`;
+     </article>`
+  );
 };
 
-export {
-  createFilmArticle
-};
+export {createFilmArticle};
